new file:   .github/workflows/tsfile.yml 	modified:   CMakeLists.txt 	modified:   include/tsfile/common/algorithm.h 	modified:   include/tsfile/common/bloomfilter.h 	modified:   include/tsfile/common/bytebuffer.h 	modified:   include/tsfile/common/concepts.h 	modified:   include/tsfile/common/hash.h 	modified:   include/tsfile/common/pack.h 	modified:   include/tsfile/model/chunk.h 	modified:   test/e2e/.settings/org.eclipse.buildship.core.prefs 	modified:   test/e2e/e2e.py 	modified:   test/model/test_bloomfilter.cc
 Untracked files:

+++ include/tsfile/model/chunk.h
@@ -146,7 +146,7 @@ class Chunk {
     const_reverse_iterator crend() const;
     /// @brief Object hash
     ///
-    uint64_t GetHashCode() const;
+    size_t GetHashCode() const;
     /// @brief Get the number of pages
     /// @return the number of pages
     ///
@@ -163,4 +163,5 @@ class Chunk {
 };
 
 }  // namespace tsfile
+
 #endif
